cmd/starlet: improve development mode
In development mode, cmd/starlet now loads bot state from a txtar file,
listens on the host:port provided in the -addr flag, and redirects to a
new /debug/bot page.

This new page is a simple JavaScript SPA that allows chatting with the
bot. It sends fake webhook requests to /telegram and uses a mocked
http.Client to intercept Telegram Bot API requests and display them on
the page. Gemini API requests still go through the normal API.